A lot of smaller fixes, but still something might be wrong with delta = 0 as sanity check with ionization2D is not quite right

+++ core/vtk/ttkMergeTreeDistanceMatrix/ttkMergeTreeDistanceMatrix.h
@@ -258,7 +258,7 @@ public:
     thresholdOfCBD_ = d;
     Modified();
   }
-  vtkGetMacro(thresholdOfCBD_, int);
+  vtkGetMacro(thresholdOfCBD_, double);
   
   bool GetThresholdOfCBD() {
     return thresholdOfCBD_;
